Fix for JIRA SB-1540
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@69388 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/align_format/align_format_util.hpp
@@ -1082,8 +1082,8 @@ public:
     ///@return:string containing template data replaced by real data
     ///
     ///<@tmplParamName@> is replaced by templParamVal
-    static string MapTemplate(string inpString,string tmplParamName,int templParamVal);
-
+    static string MapTemplate(string inpString,string tmplParamName,Int8 templParamVal);
+    
     ///Replace template tags by real data
     ///@param inpString: string containing template data
     ///@param tmplParamName:string with template tag name
